LPS-113277 Move util classes to util package
