+++ modules/apps/account/account-test/src/testIntegration/java/com/liferay/account/service/test/util/AccountEntryTestUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.account.service.test;
+package com.liferay.account.service.test.util;
 
 import com.liferay.account.constants.AccountConstants;
 import com.liferay.account.model.AccountEntry;
